Add dashboard tabs and components
