+++ src/components/dashboard/DashboardExperiences.js
@@ -1,0 +1,9 @@
+const DashboardExperiences = () => {
+  return (
+    <div className="dashboard-placeholder">
+      <h2>Experience management coming soon.</h2>
+    </div>
+  );
+};
+
+export default DashboardExperiences;
