+++ src/components/dashboard/DashboardProjects.js
@@ -1,0 +1,9 @@
+const DashboardProjects = () => {
+  return (
+    <div className="dashboard-placeholder">
+      <h2>Project management coming soon.</h2>
+    </div>
+  );
+};
+
+export default DashboardProjects;
